backend/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
boto3==1.26.
python-dotenv==0.21.1